feat(ui): Do not reset Organization store on 429 request errors
By default, `react-query` will refetch stale queries on window focus.
This can lead to many queries that can reach Sentrys API rate limit.

In the case of the organization details endpoint, an error means that the
OrganizationStore will not have an organization, which leads to
`useOrganization` throwning because organization is undefined. This ultimately
leads to a "crash" for the user (e.g. blank page w/ an error message).

This PR does not call the OrganizationStore error handler in the case of a 429
(granted the organization has not changed).

Note: I had to split up the query option creators from the hooks so that we
could mock them for the test
