+++ static/app/views/organizationLayout/index.spec.tsx
@@ -1,12 +1,17 @@
 import {OrganizationFixture} from 'sentry-fixture/organization';
 
-import {render, screen} from 'sentry-test/reactTestingLibrary';
+import {initializeOrg} from 'sentry-test/initializeOrg';
+import {render, screen, waitFor} from 'sentry-test/reactTestingLibrary';
 
+import * as bootstrapRequestOptions from 'sentry/bootstrap/bootstrapRequestOptions';
 import AlertStore from 'sentry/stores/alertStore';
 import OrganizationStore from 'sentry/stores/organizationStore';
 import PageFiltersStore from 'sentry/stores/pageFiltersStore';
 import ProjectsStore from 'sentry/stores/projectsStore';
-import OrganizationLayout from 'sentry/views/organizationLayout';
+import {queryOptions} from 'sentry/utils/queryClient';
+import App from 'sentry/views/app';
+
+import OrganizationLayout from './index';
 
 jest.mock(
   'sentry/components/sidebar',
@@ -137,5 +142,96 @@ describe('OrganizationLayout', function () {
     expect(
       await screen.findByText(/Celery workers have not checked in/)
     ).toBeInTheDocument();
+  });
+
+  describe('organizaztion details request errors', () => {
+    const organization = OrganizationFixture();
+    function SimpleChild() {
+      return <div>hello world</div>;
+    }
+
+    beforeEach(() => {
+      MockApiClient.addMockResponse({
+        url: '/organizations/',
+        body: [organization],
+      });
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/',
+        body: organization,
+      });
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/teams/',
+        body: [],
+      });
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/projects/',
+        body: [],
+      });
+      MockApiClient.addMockResponse({
+        url: '/assistant/',
+        body: [],
+      });
+    });
+
+    it('handles 429 errors', async () => {
+      const orgQueryOptions =
+        bootstrapRequestOptions.getBootstrapOrganizationQueryOptions(organization.slug);
+
+      const {router, routerProps} = initializeOrg({
+        organization,
+        router: {
+          params: {orgId: organization.slug},
+        },
+      });
+
+      const {rerender} = render(
+        <App {...routerProps}>
+          <OrganizationLayout>
+            <SimpleChild />
+          </OrganizationLayout>
+        </App>,
+        {router, deprecatedRouterMocks: true}
+      );
+      await waitFor(() => {
+        expect(
+          screen.queryByText('Loading data for your organization')
+        ).not.toBeInTheDocument();
+      });
+      expect(screen.getByText('hello world')).toBeInTheDocument();
+
+      // mock console.error to prevent it from failing jest tests since we're expecting the request error
+      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
+
+      jest
+        .spyOn(bootstrapRequestOptions, 'getBootstrapOrganizationQueryOptions')
+        .mockImplementation(() => {
+          return queryOptions({
+            ...orgQueryOptions,
+            // change queryKey to force a new query
+            queryKey: ['bootstrap-organization-429', organization.slug],
+          });
+        });
+
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/',
+        body: null,
+        statusCode: 429,
+      });
+
+      // Simulate a second API call (e.g., user focuses page after being idle, react-query queries re-fire due to expiration
+      rerender(
+        <App {...routerProps}>
+          <OrganizationLayout>
+            <SimpleChild />
+          </OrganizationLayout>
+        </App>
+      );
+
+      await waitFor(() => {
+        expect(screen.getByText('hello world')).toBeInTheDocument();
+      });
+
+      consoleErrorSpy.mockRestore();
+    });
   });
 });
